feat: ReportDetail, rm mock data
